# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/org/zones": {
      "get": {
        "operationId": "OrgController_listZones",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ZoneResponseDtoClass"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Org"
        ]
      },
      "post": {
        "operationId": "OrgController_createZone",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateZoneBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedIdDto"
                }
              }
            }
          }
        },
        "tags": [
          "Org"
        ]
      }
    },
    "/org/municipios": {
      "get": {
        "operationId": "OrgController_listMunicipios",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MunicipioResponseDtoClass"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Org"
        ]
      },
      "post": {
        "operationId": "OrgController_createMunicipio",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMunicipioBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedIdDto"
                }
              }
            }
          }
        },
        "tags": [
          "Org"
        ]
      }
    },
    "/org/users": {
      "get": {
        "operationId": "OrgController_listUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Org"
        ]
      },
      "post": {
        "operationId": "OrgController_provisionUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvisionUserBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedIdDto"
                }
              }
            }
          }
        },
        "tags": [
          "Org"
        ]
      }
    },
    "/org/coordinadores/assign": {
      "post": {
        "operationId": "OrgController_assignCoordinador",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignCoordinadorBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Coordinator assigned successfully (no body)"
          }
        },
        "tags": [
          "Org"
        ]
      }
    },
    "/org/zones/{id}": {
      "patch": {
        "operationId": "OrgController_updateZone",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateZoneBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZoneResponseDtoClass"
                }
              }
            }
          }
        },
        "tags": [
          "Org"
        ]
      },
      "delete": {
        "operationId": "OrgController_deleteZone",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Zone deleted"
          }
        },
        "tags": [
          "Org"
        ]
      }
    },
    "/org/municipios/{id}": {
      "patch": {
        "operationId": "OrgController_updateMunicipio",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMunicipioBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MunicipioResponseDtoClass"
                }
              }
            }
          }
        },
        "tags": [
          "Org"
        ]
      },
      "delete": {
        "operationId": "OrgController_deleteMunicipio",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Municipio deleted"
          }
        },
        "tags": [
          "Org"
        ]
      }
    },
    "/org/areas": {
      "get": {
        "operationId": "OrgController_listAreas",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AreaResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Org"
        ]
      },
      "post": {
        "operationId": "OrgController_createArea",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAreaBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedIdDto"
                }
              }
            }
          }
        },
        "tags": [
          "Org"
        ]
      }
    },
    "/org/areas/{id}": {
      "patch": {
        "operationId": "OrgController_updateArea",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAreaBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreaResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Org"
        ]
      },
      "delete": {
        "operationId": "OrgController_deleteArea",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Area deleted"
          }
        },
        "tags": [
          "Org"
        ]
      }
    },
...
  },
  "info": {
    "title": "FuturaGest API",
    "description": "",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ZoneResponseDtoClass": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "description": "ISO 8601 timestamp"
          },
          "updatedAt": {
            "type": "string",
            "description": "ISO 8601 timestamp — delta cursor for ?since= queries"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
      "MunicipioResponseDtoClass": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "zoneId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "description": "ISO 8601 timestamp"
          },
          "updatedAt": {
            "type": "string",
            "description": "ISO 8601 timestamp — delta cursor for ?since= queries"
          }
        },
        "required": [
          "id",
          "name",
          "zoneId",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "description": "Role enum (SYSTEM_ADMIN, GERENCIA, …)"
          },
          "mustChangePassword": {
            "type": "boolean"
          },
          "coordinatedZoneId": {
            "type": "string",
            "nullable": true,
            "description": "Zone coordinated (COORDINADOR only)"
          },
          "createdAt": {
            "type": "string",
            "description": "ISO 8601 timestamp"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "mustChangePassword",
          "coordinatedZoneId",
          "createdAt"
        ]
      },
      "AssignCoordinadorBody": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "zoneId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "userId",
          "zoneId"
        ]
      },
      "ProvisionUserBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "GERENCIA",
              "TALENTO_HUMANO",
              "LIDER_OPERATIVO"
            ]
          }
        },
        "required": [
          "email",
          "password",
          "role"
        ]
      },
      "CreatedIdDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "id"
        ]
      },
      "CreateZoneBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Unique zone name"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateZoneBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "New zone name"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateMunicipioBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Municipio name (unique within zone)"
          },
          "zoneId": {
            "type": "string",
            "format": "uuid",
            "description": "Zone this municipio belongs to"
          }
        },
        "required": [
          "name",
          "zoneId"
        ]
      },
      "UpdateMunicipioBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "New municipio name"
          },
          "zoneId": {
            "type": "string",
            "format": "uuid",
            "description": "New zone for this municipio"
          }
        }
      },
      "AreaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "horaInicio": {
            "type": "string",
            "description": "HH:MM"
          },
          "horaFin": {
            "type": "string",
            "description": "HH:MM"
          },
          "zoneId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "description": "ISO 8601 timestamp"
          },
          "updatedAt": {
            "type": "string",
            "description": "ISO 8601 timestamp"
          }
        },
        "required": [
          "id",
          "name",
          "horaInicio",
          "horaFin",
          "zoneId",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateAreaBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Area name (unique within zone)"
          },
          "horaInicio": {
            "type": "string",
            "description": "HH:MM",
            "example": "06:00"
          },
          "horaFin": {
            "type": "string",
            "description": "HH:MM",
            "example": "14:00"
          },
          "zoneId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "name",
          "horaInicio",
          "horaFin",
          "zoneId"
        ]
      },
...
    }
  }
}